+++ integrations/github/github/webhook/events.py
@@ -10,7 +10,8 @@ REPOSITORY_UPSERT_EVENTS = [
 ]
 REPOSITORY_DELETE_EVENTS = ["archived", "deleted"]
 
+
 ALL_EVENTS = REPOSITORY_UPSERT_EVENTS + REPOSITORY_DELETE_EVENTS
 
 
-WEBHOOK_CREATE_EVENTS = ["repository"]
+WEBHOOK_CREATE_EVENTS = ["repository", "deployment", "deployment_status"]